monitor-api/requirements.txt
aniso8601==1.1.0
contoml==0.32
cryptography==1.3.1
Flask==0.10.1
flask-marshmallow==0.6.2
Flask-RESTful==0.3.5
Flask-SQLAlchemy==2.1
idna==2.1
invoke==0.12.2
iso8601==0.1.11
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
marshmallow==2.6.1
psycopg2==2.6.1
py==1.4.31
pyasn1==0.1.9
pycparser==2.14
PyJWT==1.4.0
pytest==2.9.1
python-dateutil==2.5.1
pytz==2016.1
six==1.10.0
SQLAlchemy==1.0.12
timestamp==0.0.1
Werkzeug==0.11.5
